# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Inventory management system API for QwikShelf",
        "title": "QwikShelf API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "QwikShelf Support",
            "email": "support@qwikshelf.local"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/procurements": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns paginated list of all purchase orders",
                "tags": [
                    "Procurements"
                ],
                "summary": "List purchase orders",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_qwikshelf_api_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_qwikshelf_api_internal_application_dto.ProcurementResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_qwikshelf_api_pkg_response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new purchase order for a supplier with line items",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Procurements"
                ],
                "summary": "Create purchase order",
                "parameters": [
                    {
                        "description": "Purchase order details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_qwikshelf_api_internal_application_dto.CreateProcurementRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_qwikshelf_api_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_qwikshelf_api_internal_application_dto.ProcurementResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_qwikshelf_api_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_qwikshelf_api_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_qwikshelf_api_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/procurements/supplier/{supplierId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all purchase orders for a specific supplier",
                "tags": [
                    "Procurements"
                ],
                "summary": "List purchase orders by supplier",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Supplier ID",
                        "name": "supplierId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_qwikshelf_api_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_qwikshelf_api_internal_application_dto.ProcurementResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_qwikshelf_api_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/procurements/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a purchase order with all its line items and supplier/product details",
                "tags": [
                    "Procurements"
                ],
                "summary": "Get purchase order",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Procurement ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_qwikshelf_api_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_qwikshelf_api_internal_application_dto.ProcurementResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_qwikshelf_api_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/procurements/{id}/receive": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Records partial or full receipt of items in a purchase order",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Procurements"
                ],
                "summary": "Receive items",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Procurement ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Items received",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_qwikshelf_api_internal_application_dto.ReceiveProcurementItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_qwikshelf_api_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_qwikshelf_api_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_qwikshelf_api_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_qwikshelf_api_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/procurements/{id}/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates status. When set to 'received', inventory is auto-adjusted.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Procurements"
                ],
                "summary": "Update purchase order status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Procurement ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New status",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_qwikshelf_api_internal_application_dto.UpdateProcurementStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_qwikshelf_api_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_qwikshelf_api_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_qwikshelf_api_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_qwikshelf_api_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_qwikshelf_api_internal_application_dto.CreateProcurementItemRequest": {
            "type": "object",
            "required": [
                "quantity",
                "unit_cost",
                "variant_id"
            ],
            "properties": {
                "quantity": {
                    "type": "number"
                },
                "unit_cost": {
                    "type": "number"
                },
                "variant_id": {
                    "type": "integer"
                }
            }
        },
        "github_com_qwikshelf_api_internal_application_dto.CreateProcurementRequest": {
            "type": "object",
            "required": [
                "items",
                "supplier_id",
                "warehouse_id"
            ],
            "properties": {
                "expected_delivery": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/github_com_qwikshelf_api_internal_application_dto.CreateProcurementItemRequest"
                    }
                },
                "supplier_id": {
                    "type": "integer"
                },
                "warehouse_id": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_qwikshelf_api_internal_application_dto.ProcurementItemResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "line_total": {
                    "type": "number"
                },
                "quantity_ordered": {
                    "type": "number"
                },
                "quantity_received": {
                    "type": "number"
                },
                "unit_cost": {
                    "type": "number"
                },
                "variant_id": {
                    "type": "integer"
                },
                "variant_name": {
                    "type": "string"
                },
                "variant_sku": {
                    "type": "string"
                },
                "variant_unit": {
                    "type": "string"
                }
            }
        },
        "github_com_qwikshelf_api_internal_application_dto.ProcurementResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "expected_delivery": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_qwikshelf_api_internal_application_dto.ProcurementItemResponse"
                    }
                },
                "ordered_by_user_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "supplier_id": {
                    "type": "integer"
                },
                "supplier_name": {
                    "type": "string"
                },
                "total_cost": {
                    "type": "number"
                },
                "warehouse_id": {
                    "type": "integer"
                },
                "warehouse_name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_qwikshelf_api_internal_application_dto.ReceiveItemRequest": {
            "type": "object",
            "required": [
                "item_id",
                "quantity_received"
            ],
            "properties": {
                "item_id": {
                    "type": "integer"
                },
                "quantity_received": {
                    "type": "number"
                }
            }
        },
        "github_com_qwikshelf_api_internal_application_dto.ReceiveProcurementItemRequest": {
            "type": "object",
            "required": [
                "items"
            ],
            "properties": {
                "items": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/github_com_qwikshelf_api_internal_application_dto.ReceiveItemRequest"
                    }
                }
            }
        },
...
        "github_com_qwikshelf_api_internal_application_dto.UpdateProcurementStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string",
                    "enum": [
                        "pending",
                        "approved",
                        "ordered",
                        "partial",
                        "received",
                        "cancelled"
                    ]
                }
            }
        },
...
        "github_com_qwikshelf_api_pkg_response.ErrorInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_qwikshelf_api_pkg_response.Meta": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "github_com_qwikshelf_api_pkg_response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/github_com_qwikshelf_api_pkg_response.ErrorInfo"
                },
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/github_com_qwikshelf_api_pkg_response.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your JWT token with the `Bearer ` prefix, e.g. `Bearer eyJhbGci...`",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}